Changed datatype for start and end time and price per hour

+++ db/migrate/20201116223339_create_courses.rb
@@ -5,7 +5,7 @@ class CreateCourses < ActiveRecord::Migration[6.0]
       t.text :description
       t.references :teacher, null: false, foreign_key: { to_table: :users }
       t.string :category
-      t.float :price_per_hour
+      t.integer :price_per_hour
 
       t.timestamps
     end
